+++ berkeley-function-call-leaderboard/bfcl_eval/model_handler/model_style.py
@@ -4,7 +4,8 @@ from enum import Enum
 # TODO: Use all caps for enum values to maintain consistency
 class ModelStyle(Enum):
     Gorilla = "gorilla"
-    OpenAI = "gpt"
+    OpenAI_Completions = "openai-completions"
+    OpenAI_Responses = "openai-responses"
     Anthropic = "claude"
     Mistral = "mistral"
     Google = "google"